# bump gc threshold