As mentioned in https://github.com/boa-dev/boa/issues/3896, Boa's GC threshold is set too low. This was causing too many GC collections before the interpreter even started running.

We see an improvement to the 2.

```
Benchmark main: ./target/release/boa
  Time (mean ± σ): 2.2 ms ± 0.2 ms [User: 0.6 ms, System: 0.1 ms]
  Range (min … max): 2.0 ms … 3.4 ms 827 runs
```

```
Benchmark PR: ./target/release/boa
  Time (mean ± σ):       2.1 ms ±   0.1 ms    [User: 0.6 ms, System: 0.1 ms]
  Range (min … max):     1.9 ms …   2.8 ms    924 runs
```

Looking at the traces:

Before:
![image](https://github.com/boa-dev/boa/assets/936006/e2d80abc-da1e-4f7e-807c-48cd73ede886)

After (there are no GC collections in this trace)
![image](https://github.com/boa-dev/boa/assets/936006/c9c4e2a3-1208-4691-80dc-8123a83d045a)

1MB seems like a good start, we can always bump it higher or adjust after this, V8's nursery size before GC kicks in seems to be between 1-8MB,